clean bugs
