+++ engine/structure/kg_embedding/abstract_kge.py
