+++ engine/structure/kg_embedding/rotate.py
@@ -1,7 +1,7 @@
 import torch
 from torch import nn
 
-from .kge_interface import KnowledgeGraphEmbedding
+from .abstract_kge import KnowledgeGraphEmbedding
 
 
 class RotatE(KnowledgeGraphEmbedding, nn.Module):
